[FIX] id type change to Long

+++ data/src/main/java/com/nikmaram/data/dataSource/MusicDataSource.kt
@@ -5,5 +5,5 @@ import com.nikmaram.data.utility.ResultData
 
 interface MusicDataSource {
     suspend fun getMusicFiles(): ResultData<List<MusicFile>>
-    suspend fun getMusicFileById(id: Int): ResultData<MusicFile>
+    suspend fun getMusicFileById(id: Long): ResultData<MusicFile>
 }
